requirements.txt
olefile==0.44
Pillow==4.3.0
pygame==1.9.3
